backend/requirements/dev.txt
-r base.txt

# Development tools
black==24.4.2
isort==5.13.2
ruff==0.4.4
mypy==1.10.0
django-stubs==5.2.8
djangorestframework-stubs==3.16.6

# Testing
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==5.0.0
pytest-mock==3.14.0
factory-boy==3.3.0
faker==40.1.0

# Development server
werkzeug==3.1.5

# Debugging
django-debug-toolbar==6.1.0
ipython==9.8.0

# Security
bandit==1.9.2
safety==3.7.0